+++ common/src/main/java/com/st0x0ef/stellaris/client/screens/TestScreen.java
@@ -1,11 +1,7 @@
 package com.st0x0ef.stellaris.client.screens;
 
 import com.st0x0ef.stellaris.client.screens.windows.TestWindow;
-import com.st0x0ef.stellaris.common.menus.TabletMenu;
 import com.st0x0ef.stellaris.common.menus.TestMenu;
-import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.client.gui.screens.Screen;
-import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Inventory;