+++ common/src/main/java/com/st0x0ef/stellaris/common/commands/StellarisCommands.java
@@ -9,7 +9,6 @@ import com.mojang.brigadier.context.CommandContext;
 import com.st0x0ef.stellaris.Stellaris;
 import com.st0x0ef.stellaris.common.data.planets.Planet;
 import com.st0x0ef.stellaris.common.data.planets.StellarisData;
-import com.st0x0ef.stellaris.common.data.recipes.SpaceStationRecipesManager;
 import com.st0x0ef.stellaris.common.launchpads.LaunchPad;
 import com.st0x0ef.stellaris.common.launchpads.LaunchPadLauncher;
 import com.st0x0ef.stellaris.common.launchpads.LaunchPadUtils;