+++ common/src/main/java/com/st0x0ef/stellaris/common/menus/TestMenu.java
@@ -1,7 +1,6 @@
 package com.st0x0ef.stellaris.common.menus;
 import com.st0x0ef.stellaris.common.registry.MenuTypesRegistry;
 import net.minecraft.network.FriendlyByteBuf;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;