+++ common/src/main/java/com/st0x0ef/stellaris/common/network/packets/TeleportEntityToPlanetPacket.java
@@ -4,7 +4,6 @@ import com.st0x0ef.stellaris.Stellaris;
 import com.st0x0ef.stellaris.common.data.planets.Planet;
 import com.st0x0ef.stellaris.common.events.custom.PlanetSelectionServerEvents;
 import com.st0x0ef.stellaris.common.network.NetworkRegistry;
-import com.st0x0ef.stellaris.common.registry.EntityData;
 import com.st0x0ef.stellaris.common.utils.PlanetUtil;
 import com.st0x0ef.stellaris.common.utils.Utils;
 import dev.architectury.event.EventResult;
@@ -66,10 +65,8 @@ public class TeleportEntityToPlanetPacket implements CustomPacketPayload {
 
             if(rocket.getPassengers().size() == 1) {
                 List<Entity> passengers = rocket.getPassengers();
-                if (!passengers.isEmpty()) {
-                    Utils.changeDimension((Player) passengers.getFirst(), planet);
-                    player.stellaris$setPlanetMenuOpen(false, (Player) rocket.getPassengers().getFirst(), true);
-                }
+                Utils.changeDimension((Player) passengers.getFirst(), planet);
+                player.stellaris$setPlanetMenuOpen(false, (Player) passengers.getFirst(), true);
             } else {
                 Utils.changeDimensionForPlayers(rocket.getPassengers(), planet, packet.coords, true);
             }