+++ common/src/main/java/com/st0x0ef/stellaris/common/utils/PlanetUtil.java
@@ -11,8 +11,6 @@ import com.st0x0ef.stellaris.common.oxygen.GlobalOxygenManager;
 import dev.architectury.registry.menu.ExtendedMenuProvider;
 import dev.architectury.registry.menu.MenuRegistry;
 import io.netty.buffer.Unpooled;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.Component;